// src/generated/PetRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type PetRequest = { "type": "view" } | { "type": "openLauncher" } | { "type": "connectFavorite" } | { "type": "beginDrag", x: number, y: number, } | { "type": "drag", x: number, y: number, } | { "type": "endDrag" };